Mac PW Pool: Restrict software installation/updates
For whatever reason, non-admin users are permitted to install and update
software on Macs by default.  This is highly undesirable in a CI
environment, and especially so in one where the underlying resources are
shared across testing contexts.  Block this by altering system settings
to require admin access.

Further through experimentation, it was found that rosetta (allows arm64
macs to run x86_64 code) ignores the admin-required settings.  To give
pause to any users trying to run `softwareupdate`, move it out of general
reach.  This isn't a perfect solution, but should at least discourage all
simple usage.

Signed-off-by: Chris Evich <cevich@redhat.com>